@@ -3,7 +3,7 @@ name = "gurobi-machinelearning"
 version = "1.5.5dev0"
 description = "package to insert ML models in Gurobi"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = {text="Apache-2.0"}
 keywords = ["mathematical optimization", "gurobi", "scikit-learn", "pytorch", "tensorflow", "ml"]
 authors = [
@@ -21,7 +21,7 @@ classifiers = [
 
 dependencies = [
   "numpy >= 1.22.0",
-  "gurobipy >= 10.0.0",
+  "gurobipy >= 11.0.0",
   "scipy >= 1.9.3"
 ]
 
