@@ -1,4 +1,4 @@
-# Copyright © 2023 Gurobi Optimization, LLC
+# Copyright © 2023-2025 Gurobi Optimization, LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -95,7 +95,7 @@ class DecisionTreeRegressorConstr(SKgetter, AbstractTreeEstimator):
         formulations = ("leafs", "paths")
         if formulation not in formulations:
             raise ValueError(
-                "Wrong value for formulation should be one of {}.".format(formulations)
+                f"Wrong value for formulation should be one of {formulations}."
             )
         self._formulation = formulation
         SKgetter.__init__(self, predictor, input_vars)
