@@ -147,40 +147,38 @@ class TestNNFormulation(unittest.TestCase):
         )
 
         cls.X = np.array(
-            (
-                [
-                    [0.06651722, 0.42833187, -0.02818223],
-                    [1.05445173, -1.07075262, -0.17992484],
-                    [0.15634897, -0.34791215, -1.98079647],
-                    [1.45427351, 0.14404357, 0.4105985],
-                    [-0.88778575, 0.37816252, 0.15494743],
-                    [0.12898291, 0.72909056, 0.0519454],
-                    [-0.20515826, 1.49407907, 0.33367433],
-                    [0.77749036, -1.25279536, -0.4380743],
-                    [-0.38732682, 1.20237985, 1.23029068],
-                    [-1.18063218, -0.51080514, 0.3869025],
-                    [-1.16514984, 0.05616534, -0.31155253],
-                    [-0.10321885, -0.15135721, 0.95008842],
-                    [1.46935877, 1.53277921, -0.18718385],
-                    [-1.53624369, 0.46566244, 0.90082649],
-                    [-0.81314628, -0.35955316, -0.67246045],
-                    [-0.74216502, 0.8644362, 0.6536186],
-                    [-1.42001794, -1.04855297, -0.30230275],
-                    [-2.55298982, -0.85409574, 0.3130677],
-                    [0.44386323, 0.12167502, 0.76103773],
-                    [0.97873798, 0.40015721, 1.76405235],
-                    [0.04575852, -1.45436567, 2.26975462],
-                    [-0.97727788, 1.86755799, 2.2408932],
-                    [-0.36274117, -0.63432209, 0.3024719],
-                    [1.17877957, 1.89588918, 1.48825219],
-                    [-0.90729836, 0.46278226, -1.63019835],
-                    [-0.89546656, -0.21274028, -1.61389785],
-                    [-0.57884966, -0.87079715, -0.68481009],
-                    [-0.50965218, 1.9507754, -1.70627019],
-                    [-0.40178094, 0.17742614, -1.7262826],
-                    [0.40234164, -1.23482582, 1.13940068],
-                ]
-            )
+            [
+                [0.06651722, 0.42833187, -0.02818223],
+                [1.05445173, -1.07075262, -0.17992484],
+                [0.15634897, -0.34791215, -1.98079647],
+                [1.45427351, 0.14404357, 0.4105985],
+                [-0.88778575, 0.37816252, 0.15494743],
+                [0.12898291, 0.72909056, 0.0519454],
+                [-0.20515826, 1.49407907, 0.33367433],
+                [0.77749036, -1.25279536, -0.4380743],
+                [-0.38732682, 1.20237985, 1.23029068],
+                [-1.18063218, -0.51080514, 0.3869025],
+                [-1.16514984, 0.05616534, -0.31155253],
+                [-0.10321885, -0.15135721, 0.95008842],
+                [1.46935877, 1.53277921, -0.18718385],
+                [-1.53624369, 0.46566244, 0.90082649],
+                [-0.81314628, -0.35955316, -0.67246045],
+                [-0.74216502, 0.8644362, 0.6536186],
+                [-1.42001794, -1.04855297, -0.30230275],
+                [-2.55298982, -0.85409574, 0.3130677],
+                [0.44386323, 0.12167502, 0.76103773],
+                [0.97873798, 0.40015721, 1.76405235],
+                [0.04575852, -1.45436567, 2.26975462],
+                [-0.97727788, 1.86755799, 2.2408932],
+                [-0.36274117, -0.63432209, 0.3024719],
+                [1.17877957, 1.89588918, 1.48825219],
+                [-0.90729836, 0.46278226, -1.63019835],
+                [-0.89546656, -0.21274028, -1.61389785],
+                [-0.57884966, -0.87079715, -0.68481009],
+                [-0.50965218, 1.9507754, -1.70627019],
+                [-0.40178094, 0.17742614, -1.7262826],
+                [0.40234164, -1.23482582, 1.13940068],
+            ]
         )
         cls.nn = {
             "intercepts": [np.array([0, 0]), np.array([0])],
